# Fix UBSan "misaligned-pointer-use" warning on aarch64