@hebasto 2024-01-20
From https://github.com/bitcoin/bitcoin/issues/29178#issuecomment-1894168783:
> Can you update [bitcoin-core/crc32c-subtree#6](https://github.com/bitcoin-core/crc32c-subtree/pull/6) so that the commit includes a description of the problem, an explanation of the fix, and any further information. We should probably add a comment inline explaining why a memcpy is being added to that code.

Done.

cc @theuni