@luke-jr 2024-01-23
This doesn't make sense. 10 lines above, we already do the same thing - why isn't that an issue? Between then and this, data is only incremented by sizeof(uint64_t) which shouldn't break alignment...

Can't we just guarantee the data is always aligned somehow?