@fanquake 2024-01-24
Compiling on an aarch64 Fedora machine, with `gcc (GCC) 14.0.1 20240118 (Red Hat 14.0.1-0)`, master (https://github.com/bitcoin/bitcoin/commit/e69796c79c0aa202087a13ba62d9fbcc1c8754d4) and this change (43180d73db591d4a955b985620fc3bc018e37bfb) produce exactly the same assembly for `crc32c::ExtendArm64`. ASM here: https://www.diffchecker.com/e9n7HIWP/.

So if there is an actual bug, this change just makes the UBSan (false positive?) warning "go away", and the bug would still be in the binary in any case?