@hebasto 2024-02-06
As far as I understand, there are two possible cases depending on the actual compiler and the ARM hardware.

1. The code works without UB. In that case, this PR makes the intention to access data via the misaligned pointer explicit and silences an UBSan warning.

2. The code does cause UB.

Assuming that it is not feasible to test every ARM hardware and compiler combination, this change is an improvement.