@theuni 2024-02-08
> > This doesn't make sense. 10 lines above, we already do the same thing - why isn't that an issue? Between then and this, data is only incremented by sizeof(uint64_t) which shouldn't break alignment...
> 
> Agree this is strange. I have no explanation.
> 
> Maybe an instruction that assumes alignment only gets emitted in the second loop? I don't have enough compiler-fu to even hazard a guess as to the difference.

Oh 🤦

This could be the case if `length < KBYTES` (`size < 1024`) :)