@theuni 2024-02-09
@hebasto Would it be possible for you to test the above theory by reproducing the UBSan warning with either:
- A gdb breakpoint that allows you to probe for size
- An assertion at the top of the function that size < 1024
- A printf of the size at the top of the function

?