@hebasto 2024-02-09
@theuni 
> Would it be possible for you to test the above theory by reproducing the UBSan warning with either...

I have no access to the aarch64 machine at the moment. I'll try to regain it as soon as possible.