@theuni 2024-02-13
> > > This doesn't make sense. 10 lines above, we already do the same thing - why isn't that an issue? Between then and this, data is only incremented by sizeof(uint64_t) which shouldn't break alignment...
> > 
> > 
> > Agree this is strange. I have no explanation.
> 
> Oh 🤦
> 
> This could be the case if `length < KBYTES` (`size < 1024`) :)

Ok, I can confirm this is the case.

I managed to reproduce on macos, and added some assertions and printfs.

While fuzzing with `utxo_snapshot` and `utxo_total_supply`, length never grows past (by my eye) 171. And an assertion that `length < KBYTES` never fails.

So I think @hebasto and @luke-jr are both correct here. The cast is not safe and should be a memcpy. But that's the case for the `__crc32cd` call 10 lines up as well, we just happen to not hit it while fuzzing.

I can confirm that we also fall into the `if(length & 4)` and `if(length & 2)` cases, though ubsan doesn't yell about those. Maybe because they're always 4byte aligned? But regardless, I think those should get the memcpy treatment as well.

Ping @hebasto . Mind extending the memcpy treatment to the other cases as well just to be safe?

@dergoegge ping for visibility