@hebasto 2024-02-14
> Mind extending the memcpy treatment to the other cases as well just to be safe?

Thanks! Done.