@dergoegge 2024-02-15
What about these: https://github.com/bitcoin-core/crc32c-subtree/blob/4d326201585eab3cd5fbf286563d6dcb15cc0c5e/src/crc32c_arm64.cc#L115-L123

I think I agree with cory that it'd be best to cover those too.