@theuni 2024-02-15
I discussed this with @madars yesterday, who suggested using `-wcast-align=strong` to check for this. Indeed it blows up here:
```c++
crc32c/src/crc32c_arm64.cc:91:28: warning: cast from ‘const uint8_t*’ {aka ‘const unsigned char*’} to ‘uint64_t*’ {aka ‘long unsigned int*’} increases required alignment of target type [-Wcast-align]
   91 |     crc = __crc32cd(crc3, *(uint64_t *)data);
      |                            ^~~~~~~~~~~~~~~~
crc32c/src/crc32c_arm64.cc:101:27: warning: cast from ‘const uint8_t*’ {aka ‘const unsigned char*’} to ‘uint64_t*’ {aka ‘long unsigned int*’} increases required alignment of target type [-Wcast-align]
  101 |     crc = __crc32cd(crc, *(uint64_t *)data);
      |                           ^~~~~~~~~~~~~~~~
crc32c/src/crc32c_arm64.cc:107:27: warning: cast from ‘const uint8_t*’ {aka ‘const unsigned char*’} to ‘uint32_t*’ {aka ‘unsigned int*’} increases required alignment of target type [-Wcast-align]
  107 |     crc = __crc32cw(crc, *(uint32_t *)data);
      |                           ^~~~~~~~~~~~~~~~
crc32c/src/crc32c_arm64.cc:112:27: warning: cast from ‘const uint8_t*’ {aka ‘const unsigned char*’} to ‘uint16_t*’ {aka ‘short unsigned int*’} increases required alignment of target type [-Wcast-align]
  112 |     crc = __crc32ch(crc, *(uint16_t *)data);
```
I think those warnings are legitimate.

It also complains about the macros which lead to `CRC32C32BYTES`, which at first glance _also_ look unsafe.

So basically the entire accelerated arm64 path is at best iffy and at worst unsafe. I'm wondering if we should just disable it completely?