@theuni 2024-02-15
> It also complains about the macros which lead to `CRC32C32BYTES`, which at first glance _also_ look unsafe.

Sure enough. I cloned the upstream crc32c repo and built/ran tests with sanitizers on my m1 macbook:

`cmake -S . -B build -DCMAKE_CXX_FLAGS=-fsanitize="undefined" -DCRC32C_BUILD_BENCHMARKS=0 -DCMAKE_CXX_COMPILER="clang++"`

```
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /Users/cory/dev/crc32c/src/crc32c_arm64.cc:83:5 in
/Users/cory/dev/crc32c/src/crc32c_arm64.cc:83:5: runtime error: load of misaligned address 0x00016fc00d39 for type 'const uint64_t' (aka 'const unsigned long long'), which requires 8 byte alignment
0x00016fc00d39: note: pointer points here
 29 f1 bf  93 6d 4d 33 1f 11 09 07  0b 15 25 3b 57 79 a1 cf  03 3d 7d c3 0f 61 b9 17  7b e5 55 cb 47
              ^
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /Users/cory/dev/crc32c/src/crc32c_arm64.cc:83:5 in
/Users/cory/dev/crc32c/src/crc32c_arm64.cc:89:27: runtime error: load of misaligned address 0x00016fc01041 for type 'uint64_t' (aka 'unsigned long long'), which requires 8 byte alignment
0x00016fc01041: note: pointer points here
 11 09 07  0b 15 25 3b 57 79 a1 cf  03 3d 7d c3 0f 61 b9 17  7b e5 55 cb 47 c9 51 df  73 0d ad 53 ff
```

Yikes.