@hebasto 2024-02-16
Addressed all cases pointed out when using `-Wcast-align=strict`.