@theuni 2024-02-16
Nice. With this patch hacked into the upstream benchmarks, I don't see any significant difference before/after:


`cmake -S . -B build -DCRC32C_BUILD_BENCHMARKS=1 -DCMAKE_BUILD_TYPE=Release -DCMAKE_CXX_COMPILER="clang++"`
before:
```
Benchmark                                   Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------
CRC32CBenchmark/Public/256               40.1 ns         40.1 ns     17462325 bytes_per_second=5.9465G/s
CRC32CBenchmark/Public/4096               309 ns          309 ns      2261705 bytes_per_second=12.3261G/s
CRC32CBenchmark/Public/65536             3474 ns         3474 ns       201535 bytes_per_second=17.57G/s
CRC32CBenchmark/Public/1048576          54909 ns        54907 ns        12712 bytes_per_second=17.7856G/s
CRC32CBenchmark/Public/16777216        884746 ns       884746 ns          783 bytes_per_second=17.6604G/s
CRC32CBenchmark/Portable/256              106 ns          106 ns      6562849 bytes_per_second=2.24057G/s
CRC32CBenchmark/Portable/4096            1169 ns         1169 ns       598066 bytes_per_second=3.26452G/s
CRC32CBenchmark/Portable/65536          18260 ns        18260 ns        38309 bytes_per_second=3.34247G/s
CRC32CBenchmark/Portable/1048576       293165 ns       293165 ns         2388 bytes_per_second=3.3311G/s
CRC32CBenchmark/Portable/16777216     4709515 ns      4709477 ns          149 bytes_per_second=3.31778G/s
CRC32CBenchmark/ArmCRC32C/256            40.1 ns         40.1 ns     17461541 bytes_per_second=5.94759G/s
CRC32CBenchmark/ArmCRC32C/4096            309 ns          309 ns      2263021 bytes_per_second=12.3365G/s
CRC32CBenchmark/ArmCRC32C/65536          3473 ns         3473 ns       201501 bytes_per_second=17.5724G/s
CRC32CBenchmark/ArmCRC32C/1048576       54904 ns        54904 ns        12691 bytes_per_second=17.7867G/s
CRC32CBenchmark/ArmCRC32C/16777216     890130 ns       890118 ns          785 bytes_per_second=17.5538G/s
```

after:
```
Benchmark                                   Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------
CRC32CBenchmark/Public/256               40.1 ns         40.1 ns     17461105 bytes_per_second=5.94675G/s
CRC32CBenchmark/Public/4096               309 ns          309 ns      2260376 bytes_per_second=12.3257G/s
CRC32CBenchmark/Public/65536             3474 ns         3474 ns       201541 bytes_per_second=17.5681G/s
CRC32CBenchmark/Public/1048576          54887 ns        54887 ns        12730 bytes_per_second=17.7924G/s
CRC32CBenchmark/Public/16777216        891310 ns       891311 ns          785 bytes_per_second=17.5304G/s
CRC32CBenchmark/Portable/256              106 ns          106 ns      6568268 bytes_per_second=2.24044G/s
CRC32CBenchmark/Portable/4096            1169 ns         1169 ns       598066 bytes_per_second=3.26383G/s
CRC32CBenchmark/Portable/65536          18263 ns        18263 ns        38298 bytes_per_second=3.34197G/s
CRC32CBenchmark/Portable/1048576       293169 ns       293155 ns         2386 bytes_per_second=3.33122G/s
CRC32CBenchmark/Portable/16777216     4710921 ns      4710926 ns          149 bytes_per_second=3.31676G/s
CRC32CBenchmark/ArmCRC32C/256            40.1 ns         40.1 ns     17462412 bytes_per_second=5.94629G/s
CRC32CBenchmark/ArmCRC32C/4096            309 ns          309 ns      2262941 bytes_per_second=12.3297G/s
CRC32CBenchmark/ArmCRC32C/65536          3474 ns         3474 ns       201600 bytes_per_second=17.5712G/s
CRC32CBenchmark/ArmCRC32C/1048576       54918 ns        54917 ns        12715 bytes_per_second=17.7824G/s
CRC32CBenchmark/ArmCRC32C/16777216     889966 ns       889953 ns          785 bytes_per_second=17.5571G/s
```

However the sanitizer warnings are now gone.